[project] rename files to English
